refactor(frontend): organise modal tabs code
Organise the code for the model tabs into folders for each tab.
