+++ frontend-v2/src/features/model/Model.tsx
@@ -27,10 +27,10 @@ import {
 import { useForm, useFormState } from "react-hook-form";
 import { FC, useCallback, useEffect, useMemo } from "react";
 import { DynamicTabs, TabPanel } from "../../components/DynamicTabs";
-import MapVariablesTab from "./MapVariablesTab";
+import MapVariablesTab from "./variables/MapVariablesTab";
 import PKPDModelTab from "./PKPDModelTab";
 import SecondaryParametersTab from "./secondary/SecondaryParameters";
-import ParametersTab from "./ParametersTab";
+import ParametersTab from "./parameters/ParametersTab";
 import useDirty from "../../hooks/useDirty";
 import { SubPageName } from "../main/mainSlice";
 import { TableHeader } from "../../components/TableHeader";
