+++ frontend-v2/src/features/model/parameters/paramPriority.ts
@@ -1,4 +1,4 @@
-import { Variable } from "../../app/backendApi";
+import { Variable } from "../../../app/backendApi";
 
 const libraryParamOrder = [
   "CL",
