+++ frontend-v2/src/features/model/parameters/resetToSpeciesDefaults.ts
@@ -8,7 +8,7 @@ import {
   UnitRead,
   VariableRead,
   VariableUpdateApiArg,
-} from "../../app/backendApi";
+} from "../../../app/backendApi";
 import paramPriority from "./paramPriority";
 import { param_default as paramDefaults } from "./param_default";
 import {
