+++ frontend-v2/src/features/simulation/Simulations.tsx
@@ -40,9 +40,9 @@ import SimulationPlotView from "./SimulationPlotView";
 import useSimulation from "./useSimulation";
 import useSimulationInputs from "./useSimulationInputs";
 import useDirty from "../../hooks/useDirty";
-import paramPriority from "../model/paramPriority";
+import paramPriority from "../model/parameters/paramPriority";
 import { selectIsProjectShared } from "../login/loginSlice";
-import { getConstVariables } from "../model/resetToSpeciesDefaults";
+import { getConstVariables } from "../model/parameters/resetToSpeciesDefaults";
 import useSubjectGroups from "../../hooks/useSubjectGroups";
 import useExportSimulation from "./useExportSimulation";
 import { SimulationsSidePanel } from "./SimulationsSidePanel";
